+++ src/services/websocket/WebSocket.js
@@ -10,7 +10,7 @@ import { updateLobby } from '@slices/LobbySlice'
 import { addNotification } from '@slices/NotificationSlice'
 import { restartQueue, updateUser } from '@slices/UserSlice'
 
-import { match, preMatch } from '@slices/MatchmakingSlice'
+import { updateMatch, updatePreMatch } from '@slices/MatchmakingSlice'
 
 export const WSS = () => {
   const dispatch = useDispatch()
@@ -136,7 +136,7 @@ export const WSS = () => {
         break
 
       case 'matches/found':
-        dispatch(preMatch(data.payload))
+        dispatch(updatePreMatch(data.payload))
         break
 
       // ==== Old Websockets ==== //
@@ -146,11 +146,11 @@ export const WSS = () => {
         break
 
       case 'ws_preMatch':
-        dispatch(preMatch(data.payload))
+        dispatch(updatePreMatch(data.payload))
         break
 
       case 'ws_preMatchCancel':
-        dispatch(preMatch(null))
+        dispatch(updatePreMatch(null))
         break
 
       case 'ws_preMatchCancelWarn':
@@ -168,7 +168,7 @@ export const WSS = () => {
         break
 
       case 'ws_match':
-        dispatch(match(data.payload))
+        dispatch(updateMatch(data.payload))
         break
 
       default: